Ensure the apt cache is updated before installing BIND
The apt cache may be out of date, so to ensure that it is
fresh, we force the update. We also implement a try/rescue
block rather than using ignore_failure as that is a more
appropriate way to do a fallback method for Ansible 2.x
onwards.

We do not use the 'name' keyword for 'block' because that
was only made available in Ansible 2.3, and these playbooks
need to work with newton, which is Ansible 2.1.

JIRA: RI-480